+++ autonomy/test_tools/docker/tendermint.py
@@ -32,7 +32,7 @@ from docker.models.containers import Container
 
 from autonomy.constants import TENDERMINT_IMAGE_NAME, TENDERMINT_IMAGE_VERSION
 from autonomy.test_tools.docker.base import DockerImage
-from autonomy.test_tools.utils import tendermint_health_check
+from autonomy.test_tools.helpers.base import tendermint_health_check
 
 
 _TCP = "tcp://"
